Rename program item signup components
