+++ client/src/views/program-item/signup/components/lottery-signup/ProgramItemLotterySignup.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from "react-i18next";
 import styled from "styled-components";
 import { Link } from "react-router";
 import { ProgramItem } from "shared/types/models/programItem";
-import { LotterySignupForm } from "client/views/program-item/signup/components/LotterySignupForm";
+import { LotterySignupForm } from "client/views/program-item/signup/components/lottery-signup/LotterySignupForm";
 import {
   DeleteLotterySignupErrorMessage,
   submitDeleteLotterySignup,
@@ -36,7 +36,7 @@ enum ClientError {
   GROUP_TOO_BIG = "group.groupTooBigWarning",
 }
 
-export const LotterySignupProgramItem = ({
+export const ProgramItemLotterySignup = ({
   programItem,
   startTime,
   lotterySignups,
